Corrected some of the method calls, added another line to display attempts remaining (just extra, felt like it)
Changed the correctNumber variable and GuessGameMenu method (the latter was returning a value it didn't need to, so I set it to void). It was set to type int because I didn't understand why I couldn't plug in a value in the return statement in the makeNumber() method. This was resolved by changing the return value to just the variable and assigning the correctNumber variable the value from the makeNumber method.